Child Tracking System i
The Child Tracking System is an innovative safety solution designed to monitor and ensure the security of children in real-time. This system utilizes GPS for precise location tracking, GSM for communication with parents or guardians via SMS, an accelerometer to detect unusual movements such as sudden falls or running, and a buzzer for immediate audible alerts in emergency situations.